% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cossim_to_angsim.R
\name{angsim_to_cossim}
\alias{angsim_to_cossim}
\title{Convert angular similarity to cosine similarity}
\usage{
angsim_to_cossim(x)
}
\arguments{
\item{x}{vector of angular similarities}
}
\value{
similarity vector of cosine similarities
}
\description{
If an angular similarity has been calculated, then this will convery it to
cosine similarities. A cosine similarity is bound between -1 and 1, with
1 being similar and -1 being similar, 0 being orthogonal.
}
\examples{
vals <- runif(100) # 100 values between 0 and 1
angsim_to_cossim(vals)
}

