% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angular_similarity.R
\name{angular_similarity}
\alias{angular_similarity}
\title{Cosine dissimilarity / Angular similarity}
\usage{
angular_similarity(a, b)
}
\arguments{
\item{a}{Vector}

\item{b}{Vector}
}
\value{
out Angle similarity, between 0 and 1
}
\description{
The normalised angle between two vectors, bounded between 0 and 1
}
\examples{
a <- c(-1, 0)
b <- c(-1, -1)
angular_similarity(a, b)

# Vectors of length > 2
x <- c(1, 4, 2, 2.5)
y <- c(5, 200, 4, -1)
angular_similarity(x, y)
}

