% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angular_similarity0.R
\name{angular_similarity0}
\alias{angular_similarity0}
\title{Cosine dissimilarity / Angular similarity against the origin}
\usage{
angular_similarity0(a)
}
\arguments{
\item{a}{Vector}
}
\value{
out Angle similarity, between 0 and 1
}
\description{
The normalised angle between a vector and the origin, bounded between 0 and 1
}
\examples{
a <- c(-1, 0)
angular_similarity0(a)

# Vectors of length > 2
x <- c(1, 4, 2, 2.5)
angular_similarity0(x)
}

