% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_vector.R
\name{average_vector}
\alias{average_vector}
\title{Calculates and average vector}
\usage{
average_vector(x)
}
\arguments{
\item{x}{Matrix of numerical data. Columns are components of the data, row per vector}
}
\value{
Mean vector
}
\description{
Calculates an average vector from rows of multiple vectors
}
\examples{
x <- matrix(c(1,5,2,10,3,10), ncol = 2)
average_vector(x)
}

