% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_score.R
\name{b_score}
\alias{b_score}
\title{Calculates a b-score smooth for individual microtitre plates}
\usage{
b_score(data, well, plate = 96, normalise = FALSE, matrix = FALSE)
}
\arguments{
\item{data}{Vector containing numerical values to normalise}

\item{well}{Vector of well labels, i.e "A01"}

\item{plate}{Integer indicating number of wells in the complete plate (96 or 384)}

\item{normalise}{If true, will return a matrix of residuals subtracted from raw input values}

\item{matrix}{If true, will return the values in the form of a matrix.
If false, will return a dataframe of wellID and values}
}
\value{
Normalised values
}
\description{
Performs a two-way median smooth to normalise for row and column effects in
microtitre plates. Normalises each to and column to have a median of zero.
}
\note{
Need all the wells of the plate to be entered, empty wells should be listed as NA
}
\examples{
df <- data.frame(well = num_to_well(1:96),
     val = rnorm(96))
b_score(data = df$val, well = df$well)
}

