% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.R
\name{check_control}
\alias{check_control}
\title{check for no negative control}
\usage{
check_control(df, plate_id, compound = "Metadata_compound",
  neg_compound = "DMSO")
}
\arguments{
\item{df}{dataframe}

\item{plate_id}{string, col name in df that has plate ID ref}

\item{comopund}{string, col name in df that has compound ref}

\item{neg_compoudn}{string, name of negative control compound}
}
\description{
Checks number of negative control values for each plate
and will return an error if a plate contains no negative
control values
}

