% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_feature_data}
\alias{check_feature_data}
\title{check featuredata is numeric}
\usage{
check_feature_data(data, feature_cols)
}
\arguments{
\item{data}{dataframe}

\item{feature_cols}{vector of string, feature column names}
}
\description{
check featuredata is numeric
}
