% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse.R
\name{collapse}
\alias{collapse}
\title{Aggregate/collapse a grouped dataframe to an average}
\usage{
collapse(grouped_data, average = median, metadata_prefix = NULL, ...)
}
\arguments{
\item{grouped_data}{grouped dataframe}

\item{average}{average function}

\item{metadata_prefix}{string, prefix of metadata columns}

\item{...}{extra arguments to average function}
}
\description{
description
}
