% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosine_sim.R
\name{cosine_sim}
\alias{cosine_sim}
\title{Cosine similarity}
\usage{
cosine_sim(...)
}
\arguments{
\item{...}{Either a numerical matrix, or two numerical vectors of the same
length.}
}
\value{
out A matrix if given a matrix, or a single number if given two
		vectors.
}
\description{
Calculates a cosine similarity of either a matrix or between two vectors.
If given a matrix, will return a cosine-similarity matrix. If given two
vectors will calculate the cosine-similarity between the two. Specific
functions for cosine vectors or matrices can be called with
\code{cosine_sim_vector} and \code{cosine_sim_mat} respectively.
}
\examples{
# example matrix
mat <- as.matrix(iris[,1:4])
out <- cosine_sim(mat)
image(out)

# two vectors
a <- rnorm(20)
b <- rnorm(20)
cosine_sim(a, b)
}

