% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosine_sim_mat.R
\name{cosine_sim_mat}
\alias{cosine_sim_mat}
\title{Creates a cosine similarity matrix}
\usage{
cosine_sim_mat(X)
}
\arguments{
\item{X}{matrix}
}
\value{
C cosine similarity matrix
}
\description{
For a given matrix, will create a cosine similarity matrix
}
\examples{
# example data
mat <- iris[,1:4]
out <- cosine_sim_mat(mat)
image(out)
}

