% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cossim_to_angsim.R
\name{cossim_to_angsim}
\alias{cossim_to_angsim}
\title{Convert cosine similarity to angular similarity}
\usage{
cossim_to_angsim(x)
}
\arguments{
\item{x}{vector of cosine similarity values}
}
\value{
disim vector of angular similarity values
}
\description{
If a cosine similarity has been calculated by cos_sim, then this function
will convert the value to angular similarity. Angular similarity is bound
between 0 and 1, with 1 being most similar, 0 being dissimilar, 0 being
orthogonal
}
\examples{
vals <- runif(100, -1, 1) # random values
cossim_to_angsim(vals)
}

