% Generated by roxygen2 (5.0.1): do not edit by hand
% Please edit documentation in R/cv_check.R
\name{cv_check}
\alias{cv_check}
\title{Calculates the coefficient of variation across groups}
\usage{
cv_check(data, group)
}
\arguments{
\item{data}{Vector of numerical values}

\item{group}{Vector of factors with which to group the numerical values}
}
\value{
dataframe with row per group, and a column of CV values
}
\description{
Calculates the coefficient of varation across groups
}
\examples{
a <- rnorm(100, 10, 1)
b <- rnorm(100, 10, 10)
vals <- c(a, b)
grouping <- c(rep('a', 100), rep('b', 100))
df <- data.frame(vals, grouping)
cv_check(data = df$vals,
          group = df$grouping)
}

