% Generated by roxygen2 (5.0.1): do not edit by hand
% Please edit documentation in R/cv_plot.R
\name{cv_plot}
\alias{cv_plot}
\title{Plots CV values}
\usage{
cv_plot(data, group, fancy = FALSE, trend = FALSE, rotate = FALSE,
  title = "CV Plot")
}
\arguments{
\item{data}{Vector of numerical values}

\item{group}{Vector of factors to group data}

\item{fancy}{if TRUE, uses ggplot2}

\item{trend}{if TRUE adds trend line to the points, useful for time data}

\item{rotate}{if TRUE will rotate x-axis labels}

\item{title}{Title of the plot}
}
\value{
plot
}
\description{
Function to plot CV values across factors.
}
\examples{
a <- rnorm(100, 10, 1)
b <- rnorm(100, 10, 10)
vals <- c(a, b)
grouping <- c(rep('a', 100), rep('b', 100))
df <- data.frame(vals, grouping)

cv_plot(data = df$vals,
        group = df$grouping)

cv_plot(data = df$vals,
        group = df$grouping,
        fancy = TRUE,
        trend = TRUE,
        title = "Fancy CV Plot")
}

