% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_map.R
\name{dist_map}
\alias{dist_map}
\title{Plots distributions per well in a plate layout}
\usage{
dist_map(well, data, title = "")
}
\arguments{
\item{well}{vector of alphanumeric wellIDs e.g 'A01'}

\item{data}{numeric vector}

\item{title}{Title of the plot}
}
\value{
ggplot plot
}
\description{
Produces distribution plots facetted in a plate-layout format.
}

