% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing.R
\name{drop_missing}
\alias{drop_missing}
\title{drop rows/columns with missing values}
\usage{
drop_missing(data, metadata_prefix = NULL)
}
\arguments{
\item{data}{dataframe}

\item{metadata_prefix}{string, prefix for metadata columns}
}
\description{
drop rows/columns with missing values
}
