% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing.R
\name{drop_missing_cols}
\alias{drop_missing_cols}
\title{drop feature columns with missing values}
\usage{
drop_missing_cols(data, threshold, metadata_prefix = NULL)
}
\arguments{
\item{data}{dataframe}

\item{threshold}{float remove column(s) if the proportion of missing values
equals or exceeds `threshold`}

\item{metadata_prefix}{string, prefix for metadata columns}
}
\description{
drop feature columns with missing values
}
