% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing.R
\name{drop_missing_rows}
\alias{drop_missing_rows}
\title{drop rows with missing values}
\usage{
drop_missing_rows(data, threshold, metadata_prefix = NULL)
}
\arguments{
\item{data}{dataframe}

\item{threshold}{float remove row(s) if the proportion of missing values
equals or exceeds `threshold`}

\item{metadata_prefix}{string, prefix for metadata columns}
}
\description{
drop rows with missing values
}
