% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_grid_unique.R
\name{expand_grid_unique}
\alias{expand_grid_unique}
\title{Unique combination of two vectors}
\usage{
expand_grid_unique(x, y, include.equals = FALSE)
}
\arguments{
\item{x}{Vector}

\item{y}{Vector}

\item{include.equals}{if FALSE will not return a pair that is the same
element paired against itself}
}
\value{
dataframe
}
\description{
Given two vectors will find all unique pairs.
}
\examples{
x <- c('one', 'two', 'three')
y <- c('a', 'b', 'c')

expand_grid_unique(x, y)
}

