% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_correlated.R
\name{find_correlated}
\alias{find_correlated}
\title{returns feature names to keep, without highly correlated features}
\usage{
find_correlated(data, threshold)
}
\arguments{
\item{data}{data.frame}

\item{threshold}{correlation threshold}
}
\description{
returns feature names to keep, without highly correlated features
}
