% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_feature_cols}
\alias{get_feature_cols}
\title{return feature column names}
\usage{
get_feature_cols(x, ...)
}
\arguments{
\item{x}{data}

\item{...}{arguments to `get_feature_index`}
}
\description{
return feature column names
}
