% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_feature_index}
\alias{get_feature_index}
\title{return feature column indices}
\usage{
get_feature_index(x, metadata_prefix = NULL)
}
\arguments{
\item{x}{data}

\item{metadata_prefix}{string, prefix of metadata columns}
}
\description{
return feature column indices
}
