% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_featuredata}
\alias{get_featuredata}
\title{return featuredata}
\usage{
get_featuredata(x, ...)
}
\arguments{
\item{x}{data}

\item{...}{arguments to \code{get_feature_cols}}
}
\description{
return featuredata
}
