% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_image.R
\name{get_image}
\alias{get_image}
\title{Display a merged image from RGB image URLs}
\usage{
get_image(df, row_number, display = "merge", green_url = "URL_Actin",
  blue_url = "URL_DNA", red_url = "URL_HCS")
}
\arguments{
\item{df}{dataframe containing columns of interest}

\item{row_number}{row number containing image URLs of interest}

\item{display}{display options, either a merged image or separate images
for each channel}

\item{green_url}{URL for green channel}

\item{blue_url}{URL for blue channel}

\item{red_url}{URL for red channel}
}
\value{
a raster image to the display window
}
\description{
Produces an RGB merge of images from image URLs
}

