% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_metadata_cols}
\alias{get_metadata_cols}
\title{return metadata column names}
\usage{
get_metadata_cols(x, ...)
}
\arguments{
\item{x}{data}

\item{...}{arguments to `get_metadata_index`}
}
\description{
return metadata column names
}
