% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_metadata_index}
\alias{get_metadata_index}
\title{return featuredata column indices}
\usage{
get_metadata_index(x, metadata_prefix = NULL)
}
\arguments{
\item{x}{data}

\item{metadata_prefix}{string, prefix of metadata columns}
}
\description{
return featuredata column indices
}
