% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_metadata_prefix}
\alias{get_metadata_prefix}
\title{get metadata_prefix from options if is null}
\usage{
get_metadata_prefix(metadata_prefix)
}
\description{
get metadata_prefix from options if is null
}
