% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{glog}
\alias{glog}
\title{Generalised log transformation}
\usage{
glog(x, lambda = 0)
}
\arguments{
\item{x}{vector or matrix}

\item{lambda}{transformation parameter}
}
\description{
defined as \code{log(y + sqrt(y^2 + lambda)}, where lambda = 0 is a normal
log transformation
}
