% Generated by roxygen2 (5.0.1): do not edit by hand
% Please edit documentation in R/mahal_dist.R
\name{mahal_dist}
\alias{mahal_dist}
\title{Mahalanobis Distance}
\usage{
mahal_dist(x, center = "mean", na.rm = TRUE)
}
\arguments{
\item{x}{Dataframe or matrix of numerical values}

\item{center}{Method to calculate distribution center, either mean or median}

\item{na.rm}{Whether to remove missing values for centroid calculation}
}
\value{
vector of distances
}
\description{
Calculates a Mahalanobis distance of a dataframe or a matrix.
Option for robust centroid location.The Mahalanobis distance is a
measure of distance between a point and a distribution. In this
simplified case it's the difference between each point and centroid
of the distribution produced by the points.
}
\examples{
data(iris)

iris_values <- iris[,1:4]
mahal_dist(iris_values)
mahal_dist(iris_values, center = "median")
}

