% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_vector.R
\name{norm_vector}
\alias{norm_vector}
\title{Calculate the norm (or length) of a vector}
\usage{
norm_vector(x)
}
\arguments{
\item{x}{Vector}
}
\value{
Numeric, norm of the input vector
}
\description{
Calculates the length of a vector, also known as the norm if there are more
than two elements in the vector.
}
\examples{
x <- c(3, 8)
norm_vector(x)

x <- c(3, 4, 5, 2)
norm_vector(x)
}

