% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.R
\name{normalise}
\alias{normalise}
\title{normalise against negative control}
\usage{
normalise(data, compound_col, neg_control = "DMSO", method = "subtract",
  average = median, metadata_prefix = "Metadata_", ...)
}
\arguments{
\item{data}{dataframe, can be a grouped dataframe}

\item{compound_col}{name of column containing compound information}

\item{neg_control}{name of the negative control compound in `compound_col`}

\item{method}{how to normalise, either "subtract" or "divide"}

\item{average}{average function}

\item{metadata_prefix}{string, prefix of metadata columns}

\item{...}{extras arguments passed to average}
}
\description{
description
}
