% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{pca}
\alias{pca}
\title{principal component analysis}
\usage{
pca(data, metadata_prefix = NULL, n_components = NULL, ...)
}
\arguments{
\item{data}{dataframe}

\item{metadata_prefix}{string, prefix of metadata columns}

\item{n_components}{numeric, If an integer greater than or equal to 1, then 
this many principal components will be returned.
If a floating point number between 0 and 1, then the number of principal
components will be how many needed to account for that proportion of
variance in the data.
If left blank then will return all principal components which is
equal to the number of feature columns.}

\item{...}{additional arguments to \code{prcomp}}
}
\description{
replace feature columns with principal components
}
