\name{pheno_screen-package}
\alias{pheno_screen-package}
\alias{pheno_screen}
\docType{package}
\title{
\packageTitle{pheno_screen}
}
\description{
\packageDescription{pheno_screen}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{pheno_screen}
\packageIndices{pheno_screen}
~~ An overview of how to use the ~~
~~ package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{pheno_screen}

Maintainer: \packageMaintainer{pheno_screen}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard ~~
~~ keywords, one per line, from ~~
~~ file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
