\name{phenoScreen-package}
\alias{phenoScreen-package}
\alias{phenoScreen}
\docType{package}
\title{
\packageTitle{phenoScreen}
}
\description{
\packageDescription{phenoScreen}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{phenoScreen}
\packageIndices{phenoScreen}
~~ An overview of how to use the ~~
~~ package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{phenoScreen}

Maintainer: \packageMaintainer{phenoScreen}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard ~~
~~ keywords, one per line, from ~~
~~ file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
