% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_confidence.R
\name{plot_confidence}
\alias{plot_confidence}
\title{Bivariate scatter plot with a confidence boundary}
\usage{
plot_confidence(x, y, confidence = 0.9, title = "", xlab = "x",
  ylab = "y")
}
\arguments{
\item{x}{Numerical vector}

\item{y}{Numerical vector}

\item{confidence}{Confidence interval between 0 and 1, default is 0.9 i.e 90\% CI}

\item{title}{Title of plot}

\item{xlab}{x-axis label}

\item{ylab}{y-axis label}
}
\value{
plot
}
\description{
Bivariate scatter plot with a confidence boundary calculates via a 
2-dimensional kernel density estimate
}
\examples{
x <- rnorm(1000)
y <- rnorm(1000)
plot_confidence(x, y, title = "Example plot")
}

