% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.R
\name{r_normalise}
\alias{r_normalise}
\title{Robust normalisation}
\usage{
r_normalise(df, plate_id, compound = "Metadata_compound",
  neg_compound = "DMSO", ...)
}
\arguments{
\item{df}{dataframe}

\item{plate_id}{string, name of the column of plate names, name for each plate}

\item{compound}{string, name of column of compound names}

\item{neg_compound, }{string, name of the negative control to normalise against}

\item{...}{additional arguments for \code{median}}
}
\description{
Method used in the Carpenter lab. Subtract the median feature value for
each plate from the treatment feature value and divide by the median
absolute deviation.
}
\examples{
# example data
N_PLATES <- 5
wells <- rep(num_to_well(1:96), N_PLATES)
plate_id <- rep(c("plate_1", "plate_2", "plate_3", "plate_4", "plate_5"),
		each = 96)
val1 <- rnorm(96 * N_PLATES, 10, 10)
val2 <- rnorm(96 * N_PLATES, 1, 100)
comps <- c(rep("cmpd", 80), rep("DMSO", 16))
compound <- rep(comps, N_PLATES)

df <- data.frame(Metadata_well = wells,
		 Metadata_plate_id = plate_id,
		 Metadata_compound = compound,
		 val1, val2)

df_out <- r_normalise(df,
		    plate_id = "Metadata_plate_id",
		    compound = "Metadata_compound",
		    neg_compound = "DMSO")
}

