% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_correlated.R
\name{remove_correlated}
\alias{remove_correlated}
\title{remove one of pairs of highly correlated features}
\usage{
remove_correlated(data, threshold = 0.95, metadata_prefix = "Metadata")
}
\arguments{
\item{data}{data.frame}

\item{threshold}{correlation threshold}

\item{metadata_prefix}{string, prefix of metadata columns}
}
\description{
remove one of pairs of highly correlated features
}
