% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/return_features.R
\name{return_features}
\alias{return_features}
\title{Return features describing an eigenvector}
\usage{
return_features(pca_data, pca_vector, n = 2)
}
\arguments{
\item{pca_data}{a prcomp object}

\item{pca_vector}{A two element vector of a PCA i.e c(PC1, PC2)}

\item{n}{How many features returned to describe each element of
`pca_vector`}
}
\value{
vector
}
\description{
For a PCA, if given a vector, this function will return the features most
heavily weighted for that eigenvector. N.B Experimental, not tested.
}
\examples{
data(iris)
pca_iris <- prcomp(iris[,1:4])


# features for a vector pointing straight up, essentially just
# those for PC2 in a positive direction.
return_features(pca_iris, c(0,1))
}

