% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.R
\name{scale_features}
\alias{scale_features}
\title{Scale feature data}
\usage{
scale_features(df)
}
\arguments{
\item{df}{dataframe}
}
\description{
Z-score of feature data, each features scaled separately
}

