% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssmd.R
\name{ssmd}
\alias{ssmd}
\title{Strictly standardised mean difference (SSMD)}
\usage{
ssmd(a, b, verbose = TRUE)
}
\arguments{
\item{a}{Vector}

\item{b}{Vector}

\item{verbose}{if TRUE will return a description of effect size}
}
\value{
SSMD and effect description if verbose is TRUE
}
\description{
SSMD is a measure of effect size, defined at the mean divided by the
standard deviation of a difference between two random values each from
one of the two groups.
\deqn{\frac{\beta = \mu_a - \mu_b}{\sqrt{\sigma_a^2 + \sigma_b^2 - \sigma_{ab}}}}
Beta is defined as the ratio of mean to SD of the two groups.
}
\examples{
a <- rnorm(100, 1)
b <- rnorm(100, 10)
ssmd(a, b)
}

