% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta.R
\name{theta}
\alias{theta}
\title{Angle between two vectors}
\usage{
theta(a, b)
}
\arguments{
\item{a}{Vector}

\item{b}{Vector}
}
\value{
Angle in degrees
}
\description{
Calculates the angle between two vectors, output in degrees
}
\examples{
a <- c(1, 2)
b <- c(3, 1)

theta(a,b)

a <- c(1, 2)
b <- c(-1, -2)
theta(a, b)
}

