% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta0.R
\name{theta0}
\alias{theta0}
\title{Angle between a vector and the origin}
\usage{
theta0(a)
}
\arguments{
\item{a}{Vector}
}
\value{
Angle, in degrees
}
\description{
Calculates the angle (in degrees) between a vector and the origin (1,0)
}
\examples{
a <- c(1, 0)
theta0(a)

a <- c(-1, 0)
theta0(a)

a <- c(-1, -1)
theta0(a)
}

