% Generated by roxygen2 (5.0.1): do not edit by hand
% Please edit documentation in R/z_factor.R
\name{z_factor}
\alias{z_factor}
\title{Calculates a Z-factor for two distributions}
\usage{
z_factor(positive, negative)
}
\arguments{
\item{positive}{Vector}

\item{negative}{Vector}
}
\value{
z-factor
}
\description{
Calculates a Z-factor of Z-prime for two distributions, used to assess the
separation between positive and negative controls in high-throughput screens.
A value > 0.5 is indicative of a strong assay.
}
\examples{
x <- rnorm(100, 100)
y <- rnorm(100, 10)
z_factor(x, y)
}

