% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_grid.R
\name{z_grid__}
\alias{z_grid__}
\title{Z_grid, but prettier}
\usage{
z_grid__(data, well, plate_id, ncols = 2, plate = 96, title = "",
  option = "D", each = FALSE)
}
\arguments{
\item{data}{Numerical values to be plotted}

\item{well}{Vector of well identifiers e.g "A01"}

\item{plate_id}{Vector of plate identifiers e.g "Plate_1"}

\item{ncols}{Number of columns to display multiple heatmaps}

\item{plate}{Number of wells in complete plate (96 or 384)}

\item{title}{Title of plot}

\item{option}{viridis colour option, 'A', 'B', 'C', or 'D'}

\item{each}{boolean, if true scales each plate individually, if false will
scale the pooled values of \code{data}}
}
\description{
z_grid using ggplot dark theme and viridis colour map
}
\examples{
df01 <- data.frame(well = num_to_well(1:96),
  vals = rnorm(96),
  plate = 1)

df02 <- data.frame(well = num_to_well(1:96),
  vals = rnorm(96),
  plate = 2)

df <- rbind(df01, df02)

z_grid__(data = df$vals,
       well = df$well,
       plate_id = df$plate,
       plate = 96,
       title = "Plot Title")
}

