% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_map.R
\name{z_map__}
\alias{z_map__}
\title{Pretty version of z_map}
\usage{
z_map__(data, well, plate = 96, title = "", option = "D")
}
\arguments{
\item{data}{Numerical values to be plotted}

\item{well}{Vector of well identifiers e.g "A01"}

\item{plate}{Number of wells in complete plate (96 or 384)}

\item{title}{Title of plot}

\item{option}{viridis colour option, 'A', 'B', 'C', or 'D'}
}
\value{
ggplot plot
}
\description{
z_map with dark theme and viridis colour map
}
\examples{
df <- data.frame(vals = rnorm(1:384),
  well = num_to_well(1:384, plate = 384))

z_map__(data = df$vals,
      well = df$well,
      plate = 384,
      title = "Title of plot")
}

