% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{z_score}
\alias{z_score}
\title{z_score}
\usage{
z_score(x, ...)
}
\arguments{
\item{x}{numeric vector}

\item{...}{additional arguments to mean and sd, e.g na.rm = TRUE}
}
\description{
standardise features to mean of 0 and unit variance
}
