% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_map.R
\name{dist_map}
\alias{dist_map}
\title{Plots distributions per well in a plate layout}
\usage{
dist_map(data, feature, row, column, title = "")
}
\arguments{
\item{data}{Dataframe}

\item{feature}{Numerical values indicating column index which contains the values
to produce the distribution.}

\item{row}{Number indicating the row within a plate}

\item{column}{Number inficating the column within a plate}

\item{title}{Title of the plot}
}
\value{
ggplot plot
}
\description{
Produces distribution plots facetted in a plate-layout format. Requires
columns within a dataframe labelled 'row' and 'column' indicating the plate
row and column, these can be produced with \code{num_to_well}.
}

