% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.R
\name{get_featuredata}
\alias{get_featuredata}
\title{Get featuredata columns}
\usage{
get_featuredata(x)
}
\arguments{
\item{x}{dataframe}
}
\description{
Identifies columns in a dataframe that are not labelled Metadata
}

