% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing.R
\name{impute_missing}
\alias{impute_missing}
\title{impute missing values}
\usage{
impute_missing(data, metadata_prefix = NULL)
}
\arguments{
\item{data}{dataframe}

\item{metadata_prefix}{string, prefix for metadata columns}
}
\description{
impute missing values
}
