% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise.R
\name{normalise}
\alias{normalise}
\title{Normalise per plate against negative control}
\usage{
normalise(df, plate_id, compound = "Metadata_compound",
  neg_compound = "DMSO")
}
\arguments{
\item{df}{dataframe}

\item{plate_id}{string, name of the column of plate names, name for each plate}

\item{compound}{string, name of column of compound names}

\item{neg_compound, }{name of the negative control to normalise against}
}
\description{
This function will normalise multivariate data on a plate-by-plate basis
by dividing each feature by the median of the negative control for
that feature. Note: Assumes any metadata columns are prefixed with
\code{metadata_prefix} , and everything else is numerical featuredata
}

