\name{phenoScreen-package}
\alias{phenoScreen-package}
\alias{phenoScreen}
\docType{package}
\title{
\packageTitle{phenoScreen}
}
\description{
\packageDescription{phenoScreen}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{phenoScreen}
\packageIndices{phenoScreen}

}
\author{
\packageAuthor{phenoScreen}

Maintainer: \packageMaintainer{phenoScreen}
}
\references{}
\keyword{}
\seealso{}
\examples{}
