% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{scale_features}
\alias{scale_features}
\title{scale feature columns}
\usage{
scale_features(data, metadata_prefix = NULL, ...)
}
\arguments{
\item{data}{dataframe}

\item{metadata_prefix}{string, prefix of metadata columns}

\item{...}{additional arguments to `z_score`}
}
\description{
scale each feature column by z-score
}
